Refactor: Move favorite logic to ArticleCard component
