+++ pages/index.js
@@ -2,7 +2,10 @@ import ArticleList from "@/components/ArticleList/ArticleList";
 import Navigation from "@/components/Navigation/Navigation";
 import useSWR from "swr";
 
+
 export default function Home() {
+
+
   const { data, error } = useSWR("https://www.tagesschau.de/api2/news");
 
   if (error) {
@@ -16,10 +19,11 @@ export default function Home() {
     (item) =>
       item.type === "story" && item.teaserImage?.imageVariants?.["16x9-256"]
   );
+
   return (
     <>
       <Navigation />
-      <ArticleList articles={stories} />
+      <ArticleList articles={stories}  />
     </>
   );
 }
